added post authorization

+++ app-post/backend/routes/posts.js
@@ -81,9 +81,13 @@ router.get('/:id', (req, res,next) => {
 
 router.delete('/:id',checkAuth,(req,res,next)=>{
     console.log(req.params.id);
-    Post.deleteOne({_id:req.params.id}).then(result=>{
+    Post.deleteOne({_id:req.params.id, creator:req.userData.userId}).then(result=>{
         console.log(result);
-        res.status(200).json({ message:'Post Deleted' });
+        if(result.n>0){
+            res.status(200).json({ message: "Update successful!" });
+          }else{
+            res.status(401).json({ message: "Not Authorised" });
+        }
     })
 });
 
@@ -121,8 +125,12 @@ router.put(
         content: req.body.content,
         imagePath: imagePath
       };
-      Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({ message: "Update successful!" });
+      Post.updateOne({ _id: req.params.id,creator:req.userData.userId }, post).then(result => {
+          if(result.nModified>0){
+            res.status(200).json({ message: "Update successful!" });
+          }else{
+            res.status(401).json({ message: "Not Authorised" });
+          }
       });
     }
   );
